New black (#20)
Seems to work pretty well, example PR made here:
https://github.com/SuffolkLITLab/docassemble-AssemblyLine/pull/804.

It also shows neat little summaries when it runs. Might be more useful
with a diff, but can't get black to output that and actually format the
files at the same time. Eh.

![Screenshot from 2023-12-12
11-47-57](https://github.com/SuffolkLITLab/ALActions/assets/6252212/9f980143-6a1d-4df6-88c5-1dd3dbfc1abf)

It will now also fail the action when the files aren't formatted, so
you're less likely to merge a PR thinking that everything was formatted,
not realizing that a PR had been made that you need to merge (something
I've done a lot).

Fixes #19