Start using hall monitory for efile uptime
With the addition of the "homepage" check type in hall monitor, it's feature 
equivalent to what we're doing in the efile server checks.[^1]

Bumped up the check time to 3 times a day (during business hours).

Fix #39

[^1]: we aren't parsing the JSON response of the about page and confirming
that there's a version anymore, but TBH that should be fine. That page should
only return a 200 if it's able to respond with a correct JSON payload.