Don't try to use any Jinja functions when checking names
`validate_with_stubbed_filters` is mainly to check that the filters
used really exist and aren't spelled wrong (from IRL communictaion).

In the AssemblyLine, there is a template that uses the `round` filter, which
is builtin to docassemble, isn't overriden by it. Since the `builtin_jinja_filters`
are all of the filters that are overriden in docassemble, `round` was still
the normal jinja function. Since `max` was a null function, it didn't change the input
list into an output single int, so the `round` filter failed with an invalid input type.

(Example of this failure: https://github.com/SuffolkLITLab/docassemble-AssemblyLine/actions/runs/19549536630/job/55977069996?pr=990)

To fix, we just need to add a `null_func` for `round`. Either everything is
a `null_func`, or nothing is. Making nothing be a `null_func` would require dependencies
on docassemble to have the same behavior, so we just add the null func.

To help find errors in here, I added `traceback.format_exc()` to print out the
stack traces of the exceptions, which are really necessary to understanding what's happening
(the "must be real number, not tuple" wasn't very helpful).